Address review: single availability read, footer test, dedupe receiver test mocks
